Extract codebase awareness scanner adapters
