+++ scripts/codebase_awareness/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Language adapters for the shallow codebase awareness scanner."""
